directions removed from last slide
