+++ code/src/components/ContentContainer.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-named-as-default */
 /* eslint-disable no-nested-ternary */
 /* eslint-disable import/no-named-as-default-member */
 import React from 'react'
